+++ src/Slices/AgentProcess/Core/Model.ts
