+++ src/UI/Routing/PrimaryRouteManager.ts
@@ -13,7 +13,7 @@ import {
   RouteMatch,
   Crumb,
 } from "@/Core";
-import { AgentProcess } from "@S/Agent";
+import { AgentProcess } from "@S/AgentProcess";
 import { Notification } from "@S/Notification";
 import { Resource } from "@S/Resource";
 import { paths } from "./Paths";
